+++ CookingSkill/Mod.cs
@@ -27,11 +27,9 @@ namespace CookingSkill
 
         public static double GetNoConsumeChance()
         {
-            if (Game1.player.HasCustomProfession(Skill.ProfessionConservation))
-                return 0.15;
-            else
-                return 0;
-
+            return Game1.player.HasCustomProfession(Skill.ProfessionConservation)
+                ? 0.15
+                : 0;
         }
 
         // Modifies the item based on professions and stuff
@@ -48,14 +46,10 @@ namespace CookingSkill
                 obj.Edibility = (int)(obj.Edibility * Mod.GetEdibilityMultiplier());
 
                 if (Game1.player.HasCustomProfession(Skill.ProfessionSellPrice))
-                {
                     obj.Price = (int)(obj.Price * 1.2);
-                }
 
                 if (Game1.player.HasCustomProfession(Skill.ProfessionSilver))
-                {
-                    obj.Quality = 1;
-                }
+                    obj.Quality = SObject.medQuality;
 
                 ConsumedItem[] used;
                 try
@@ -70,22 +64,23 @@ namespace CookingSkill
                 }
 
                 int total = 0;
-                foreach (ConsumedItem ingr in used)
-                    total += ingr.Amount;
+                foreach (ConsumedItem ingredient in used)
+                    total += ingredient.Amount;
 
-                for (int iq = 1; iq <= SObject.bestQuality; ++iq)
+                for (int quality = 1; quality <= SObject.bestQuality; ++quality)
                 {
-                    if (iq == 3) continue; // Not a real quality
+                    if (quality == 3)
+                        continue; // not a real quality
 
                     double chance = 0;
-                    foreach (ConsumedItem ingr in used)
+                    foreach (ConsumedItem ingredient in used)
                     {
-                        if (ingr.Item.Quality >= iq)
-                            chance += (1.0 / total) * ingr.Amount;
+                        if (ingredient.Item.Quality >= quality)
+                            chance += (1.0 / total) * ingredient.Amount;
                     }
 
                     if (rand.NextDouble() < chance)
-                        obj.Quality = iq;
+                        obj.Quality = quality;
                 }
 
                 return rand.NextDouble() >= Mod.GetNoConsumeChance();
@@ -120,133 +115,165 @@ namespace CookingSkill
 
         private void OnItemEaten(object sender, EventArgs e)
         {
-            SObject obj = Game1.player.itemToEat as SObject;
-            string[] info = Game1.objectInformation[obj.ParentSheetIndex].Split('/');
-            string[] buffData = ((Convert.ToInt32(info[2]) > 0 && info.Length > 7) ? info[7].Split(' ') : null);
+            // get object eaten
+            if (Game1.player.itemToEat is not SObject { Category: SObject.CookingCategory } obj || !Game1.objectInformation.TryGetValue(obj.ParentSheetIndex, out string rawObjData))
+                return;
+            string[] objFields = rawObjData.Split('/');
+            bool isDrink = objFields.GetOrDefault(SObject.objectInfoMiscIndex) == "drink";
 
-            if (buffData != null)
+            // get buff data
+            Buff oldBuff = isDrink ? Game1.buffsDisplay.drink : Game1.buffsDisplay.food;
+            Buff curBuff = this.CreateBuffFromObjectField(objFields);
+            if (oldBuff != null && curBuff != null && oldBuff.buffAttributes.SequenceEqual(curBuff.buffAttributes) && oldBuff != this.LastDrink)
             {
-                bool allZero = true;
-                foreach (string bd in buffData)
+                // Now that we know that this is the original buff, we can buff the buff.
+                Log.Trace("Buffing buff");
+                int[] newAttr = (int[])curBuff.buffAttributes.Clone();
+                if (Game1.player.HasCustomProfession(Skill.ProfessionBuffLevel))
                 {
-                    allZero = (allZero && bd == "0");
-                }
-                if (allZero) buffData = null;
-            }
-
-            if (info[3] == "Cooking -7")
-            {
-                // Need to make sure this is the original buff first.
-                // So it doesn't get rebuffed from eating a buff food -> non buff food -> buff food or something
-                Buff oldBuff = (info[6] == "drink" ? Game1.buffsDisplay.drink : Game1.buffsDisplay.food);
-                Buff thisBuff;
-                if (info[6] == "drink")
-                    thisBuff = buffData == null ? null : new Buff(Convert.ToInt32(buffData[0]), Convert.ToInt32(buffData[1]), Convert.ToInt32(buffData[2]), Convert.ToInt32(buffData[3]), Convert.ToInt32(buffData[4]), Convert.ToInt32(buffData[5]), Convert.ToInt32(buffData[6]), Convert.ToInt32(buffData[7]), Convert.ToInt32(buffData[8]), Convert.ToInt32(buffData[9]), Convert.ToInt32(buffData[10]), (buffData.Length > 10) ? Convert.ToInt32(buffData[10]) : 0, (info.Length > 8) ? Convert.ToInt32(info[8]) : -1, info[0], info[4]);
-                else
-                    thisBuff = buffData == null ? null : new Buff(Convert.ToInt32(buffData[0]), Convert.ToInt32(buffData[1]), Convert.ToInt32(buffData[2]), Convert.ToInt32(buffData[3]), Convert.ToInt32(buffData[4]), Convert.ToInt32(buffData[5]), Convert.ToInt32(buffData[6]), Convert.ToInt32(buffData[7]), Convert.ToInt32(buffData[8]), Convert.ToInt32(buffData[9]), Convert.ToInt32(buffData[10]), (buffData.Length > 11) ? Convert.ToInt32(buffData[11]) : 0, (info.Length > 8) ? Convert.ToInt32(info[8]) : -1, info[0], info[4]);
-                int[] oldAttr = oldBuff?.buffAttributes;
-                int[] thisAttr = thisBuff?.buffAttributes;
-                Log.Trace("Ate something: " + obj + " " + Game1.objectInformation[obj.ParentSheetIndex] + " " + buffData + " " + oldBuff + " " + thisBuff + " " + oldAttr + " " + thisAttr);
-                if (oldBuff != null && thisBuff != null && oldAttr.SequenceEqual(thisAttr) &&
-                     ((info[6] == "drink" && oldBuff != this.LastDrink) || (info[6] != "drink" && oldBuff != this.LastDrink)))
-                {
-                    // Now that we know that this is the original buff, we can buff the buff.
-                    Log.Trace("Buffing buff");
-                    int[] newAttr = (int[])thisAttr.Clone();
-                    if (Game1.player.HasCustomProfession(Skill.ProfessionBuffLevel))
+                    for (int id = 0; id < newAttr.Length; ++id)
                     {
-                        for (int i = 0; i < thisAttr.Length; ++i)
-                        {
-                            if (newAttr[i] <= 0)
-                                continue;
+                        if (newAttr[id] <= 0)
+                            continue;
 
-                            if (i == 7 || i == 8)
-                                newAttr[i] = (int)(newAttr[i] * 1.2);
-                            else
-                                newAttr[i]++;
-                        }
-                    }
-
-                    int newTime = (info.Length > 8) ? Convert.ToInt32(info[8]) : -1;
-                    if (newTime != -1 && Game1.player.HasCustomProfession(Skill.ProfessionBuffTime))
-                    {
-                        newTime = (int)(newTime * 1.25);
-                    }
-
-                    Buff newBuff = new Buff(newAttr[0], newAttr[1], newAttr[2], newAttr[3], newAttr[4], newAttr[5], newAttr[6], newAttr[7], newAttr[8], newAttr[9], newAttr[10], newAttr[11], newTime, info[0], info[4])
-                    {
-                        millisecondsDuration = newTime / 10 * 7000
-                    };
-                    // ^ The vanilla code decreases the duration based on the time of day.
-                    // This is fine normally, since it ends as the day ends.
-                    // However if you have something like TimeSpeed it just means it won't
-                    // last as long later if eaten later in the day.
-
-                    if (info[6] == "drink")
-                    {
-                        Game1.buffsDisplay.drink.removeBuff();
-                        Game1.buffsDisplay.drink = newBuff;
-                        Game1.buffsDisplay.drink.addBuff();
-                        this.LastDrink = newBuff;
-                    }
-                    else
-                    {
-                        Game1.buffsDisplay.food.removeBuff();
-                        Game1.buffsDisplay.food = newBuff;
-                        Game1.buffsDisplay.food.addBuff();
-                    }
-                    Game1.buffsDisplay.syncIcons();
-                }
-                else if (thisBuff == null && Game1.player.HasCustomProfession(Skill.ProfessionBuffPlain))
-                {
-                    Log.Trace("Buffing plain");
-                    Random rand = new Random();
-                    int[] newAttr = new int[12];
-                    int count = 1 + Math.Min(obj.Edibility / 30, 3);
-                    for (int i = 0; i < count; ++i)
-                    {
-                        int attr = rand.Next(10);
-                        if (attr >= 3) ++attr; // 3 unused?
-                        if (attr >= 6) ++attr; // 6 is crafting speed, unused?
-
-                        int amt = 1;
-                        if (attr == 7 || attr == 8)
-                            amt = 25 + rand.Next(4) * 5;
+                        if (id is Buff.maxStamina or Buff.magneticRadius)
+                            newAttr[id] = (int)(newAttr[id] * 1.2);
                         else
-                        {
-                            // 36% (assuming I used this probability calculator right) chance for a buff to be level 2
-                            // 4% chance for it to be 3
-                            if (rand.NextDouble() < 0.2)
-                                ++amt;
-                            if (rand.NextDouble() < 0.2)
-                                ++amt;
-                        }
-                        newAttr[attr] += amt;
+                            newAttr[id]++;
                     }
+                }
 
-                    int newTime = 120 + obj.Edibility / 10 * 30;
+                float minutesDuration = curBuff.millisecondsDuration / 7000f * 10f;
+                if (Game1.player.HasCustomProfession(Skill.ProfessionBuffTime))
+                    minutesDuration *= 1.25f;
 
-                    Buff newBuff = new Buff(newAttr[0], newAttr[1], newAttr[2], newAttr[3], newAttr[4], newAttr[5], newAttr[6], newAttr[7], newAttr[8], newAttr[9], newAttr[10], newAttr[11], newTime, info[0], info[4])
-                    {
-                        millisecondsDuration = newTime / 10 * 7000
-                    };
+                Buff newBuff = this.CreateBuff(newAttr, (int)minutesDuration, objFields);
+                this.ReplaceBuff(newBuff, isDrink);
+            }
+            else if (curBuff == null && Game1.player.HasCustomProfession(Skill.ProfessionBuffPlain))
+            {
+                Log.Trace("Buffing plain");
+                Random rand = new();
+                int[] newAttr = new int[12];
+                int count = 1 + Math.Min(obj.Edibility / 30, 3);
+                for (int i = 0; i < count; ++i)
+                {
+                    int attr = rand.Next(10);
+                    if (attr >= 3)
+                        ++attr; // 3 unused?
+                    if (attr >= Buff.speed)
+                        ++attr; // unused?
 
-                    if (info[6] == "drink")
-                    {
-                        Game1.buffsDisplay.drink?.removeBuff();
-                        Game1.buffsDisplay.drink = newBuff;
-                        Game1.buffsDisplay.drink.addBuff();
-                        this.LastDrink = newBuff;
-                    }
+                    int amt = 1;
+                    if (attr is Buff.maxStamina or Buff.magneticRadius)
+                        amt = 25 + rand.Next(4) * 5;
                     else
                     {
-                        Game1.buffsDisplay.drink?.removeBuff();
-                        Game1.buffsDisplay.drink = newBuff;
-                        Game1.buffsDisplay.drink.addBuff();
+                        // 36% (assuming I used this probability calculator right) chance for a buff to be level 2
+                        // 4% chance for it to be 3
+                        if (rand.NextDouble() < 0.2)
+                            ++amt;
+                        if (rand.NextDouble() < 0.2)
+                            ++amt;
                     }
-                    Game1.buffsDisplay.syncIcons();
+                    newAttr[attr] += amt;
                 }
+
+                int newTime = 120 + obj.Edibility / 10 * 30;
+
+                Buff newBuff = this.CreateBuff(newAttr, newTime, objFields);
+                this.ReplaceBuff(newBuff, isDrink);
             }
+        }
+
+        /// <summary>Create a buff instance.</summary>
+        /// <param name="attr">The buff attributes.</param>
+        /// <param name="minutesDuration">The buff duration in minutes.</param>
+        /// <param name="objectFields">The raw object fields from <see cref="Game1.objectInformation"/>.</param>
+        private Buff CreateBuff(int[] attr, int minutesDuration, string[] objectFields)
+        {
+            return new(
+                farming: attr.GetOrDefault(Buff.farming),
+                fishing: attr.GetOrDefault(Buff.fishing),
+                mining: attr.GetOrDefault(Buff.mining),
+                digging: attr.GetOrDefault(3),
+                luck: attr.GetOrDefault(Buff.luck),
+                foraging: attr.GetOrDefault(Buff.foraging),
+                crafting: attr.GetOrDefault(Buff.crafting),
+                maxStamina: attr.GetOrDefault(Buff.maxStamina),
+                magneticRadius: attr.GetOrDefault(Buff.magneticRadius),
+                speed: attr.GetOrDefault(Buff.speed),
+                defense: attr.GetOrDefault(Buff.defense),
+                attack: attr.GetOrDefault(Buff.attack),
+                minutesDuration: minutesDuration,
+                source: objectFields.GetOrDefault(SObject.objectInfoNameIndex),
+                displaySource: objectFields.GetOrDefault(SObject.objectInfoDisplayNameIndex)
+            );
+        }
+
+        /// <summary>Create a buff instance from the raw buff data in a <see cref="Game1.objectInformation"/> entry, if valid.</summary>
+        /// <param name="fields">The object fields.</param>
+        private Buff CreateBuffFromObjectField(string[] fields)
+        {
+            // get object info
+            int edibility = Convert.ToInt32(fields.GetOrDefault(SObject.objectInfoEdibilityIndex));
+            string name = fields.GetOrDefault(SObject.objectInfoNameIndex);
+            string displayName = fields.GetOrDefault(SObject.objectInfoDisplayNameIndex);
+
+            // ignore if item doesn't provide a buff
+            if (edibility < 0 || fields.Length <= SObject.objectInfoBuffTypesIndex)
+                return null;
+
+            // get buff duration
+            if (!fields.TryGetIndex(SObject.objectInfoBuffDurationIndex, out string rawDuration) || !int.TryParse(rawDuration, out int minutesDuration))
+                minutesDuration = 0;
+
+            // get buff fields
+            string[] attr = fields[SObject.objectInfoBuffTypesIndex].Split(' ');
+            if (attr.All(val => val == "0"))
+                return null;
+
+            // parse buff
+            int GetAttr(int index) => attr.TryGetIndex(index, out string raw) && int.TryParse(raw, out int value) ? value : 0;
+            return new Buff(
+                farming: GetAttr(Buff.farming),
+                fishing: GetAttr(Buff.fishing),
+                mining: GetAttr(Buff.mining),
+                digging: GetAttr(3),
+                luck: GetAttr(Buff.luck),
+                foraging: GetAttr(Buff.foraging),
+                crafting: GetAttr(Buff.crafting),
+                maxStamina: GetAttr(Buff.maxStamina),
+                magneticRadius: GetAttr(Buff.magneticRadius),
+                speed: GetAttr(Buff.speed),
+                defense: GetAttr(Buff.defense),
+                attack: GetAttr(Buff.attack),
+                minutesDuration: minutesDuration,
+                source: name,
+                displaySource: displayName
+            );
+        }
+
+        /// <summary>Replace the current food or drink buff.</summary>
+        /// <param name="newBuff">The new buff to set.</param>
+        /// <param name="isDrink">Whether the buff is a drink buff; else it's a food buff.</param>
+        private void ReplaceBuff(Buff newBuff, bool isDrink)
+        {
+            if (isDrink)
+            {
+                Game1.buffsDisplay.drink?.removeBuff();
+                Game1.buffsDisplay.drink = newBuff;
+                Game1.buffsDisplay.drink.addBuff();
+                this.LastDrink = newBuff;
+            }
+            else
+            {
+                Game1.buffsDisplay.food?.removeBuff();
+                Game1.buffsDisplay.food = newBuff;
+                Game1.buffsDisplay.food.addBuff();
+            }
+
+            Game1.buffsDisplay.syncIcons();
         }
     }
 }