+++ SpaceShared/DiagnosticMessages.cs
@@ -8,6 +8,9 @@ namespace SpaceShared
         /// <summary>The code was copied from the decompiled game, so we should avoid refactoring to simplify future updates.</summary>
         public const string CopiedFromGameCode = "The code was copied from the decompiled game, so we should avoid refactoring to simplify future updates.";
 
+        /// <summary>The class or member is used via reflection.</summary>
+        public const string IsUsedViaReflection = "The class or member is used via reflection.";
+
         /// <summary>The class is part of a mod's API, so we can't make breaking changes.</summary>
         public const string IsPublicApi = "The class is part of a mod's API, so we can't make breaking changes.";