+++ CookingSkill/docs/release-notes.md
@@ -1,6 +1,14 @@
 ﻿﻿[← back to readme](README.md)
 
 # Release notes
+## 1.3.1
+Released 11 July 2021 for Stardew Valley 1.5. Updated by Pathoschild.
+
+* Fixed multiplied recipes in Cooking SKill 1.3.0.
+* Fixed Secret Spices profession always replacing drink buffs.
+* Fixed Intense Flavors profession applying drink defense bonus to attack.
+* Internal refactoring.
+
 ## 1.3.0
 Released 10 July 2021 for Stardew Valley 1.5. Updated by Pathoschild.